Cache optimization and bug fixes.
Fixed Xcode 5.2 bugs in PXPath.m and PXParcerBase.m modules.

The styleChangable property is added to PXStyleable protocol. This
property added to all classes that support the PXStyleable protocol.
The property is used to set a flag to objects in which style is being
set programmatically (depending on a program's settings)
Accordingly, the support of this property is added to 'PXStyleTreeInfo'
and 'PXStyleInfo' modules. In case if an object with 'styleChangeble'
property is encountered, style for this object is being read rather
then used from cache.

Into stylerByState_ dictionary (PXStyleInfo class) class names are put
(rather than id<PXStyler>). The dictionary is only needed for locating
PXStyler correspondence, so in case if a styler doesn't use
sharedInstance, then 'if ([activeStylers
containsObject:currentStyler])' condition returns 'NO' because objects
are different.

In pxStyleChildren method of PXUICollectionViewCell and
PXUITableViewCell classes the string
 ' return [styleChildren arrayByAddingObjectsFromArray:self.subviews];'
is replaced with
' return styleChildren;'
contentView contains all the objects in any case, so if the cache is
used the style is applied two times.